+++ packages/saber-plugin-image/lib/saber-browser.js
@@ -1,16 +1,28 @@
-import VueLazyload from 'vue-lazyload'
+import lozad from 'lozad'
 import styles from './styles.module.css'
 
 export default ({ Vue }) => {
-  const options = Object.assign(
-    __SABER_IMAGE_OPTIONS__ || {}, // eslint-disable-line no-undef
-    { lazyComponent: true }
-  )
-
-  Vue.use(VueLazyload, options)
+  const options = __SABER_IMAGE_OPTIONS__ // eslint-disable-line no-undef
+  const blank =
+    'data:image/gif;base64,R0lGODdhAQABAPAAAMPDwwAAACwAAAAAAQABAAACAkQBADs='
 
   Vue.component('saber-image', {
     props: ['src', 'lazy'],
+    mounted() {
+      const { $el } = this
+
+      if ($el.dataset.src || $el.dataset.srcset) {
+        lozad($el, {
+          loaded(el) {
+            el.addEventListener(
+              'load',
+              () => el.setAttribute('data-lazy-loaded', ''),
+              { once: true }
+            )
+          }
+        }).observe()
+      }
+    },
     render(h) {
       const lazy = Object.assign(
         options,
@@ -28,51 +40,30 @@ export default ({ Vue }) => {
           const { src } = this
 
           return h('img', {
-            attrs: $attrs,
-            directives: [
-              {
-                name: 'lazy',
-                value: {
-                  src
-                }
-              }
-            ]
+            attrs: {
+              ...$attrs,
+              src,
+              srcset: blank,
+              'data-srcset': src
+            }
           })
         }
 
-        const { width, height, src, srcSet, placeholder } = this.src
+        const { width, src, srcSet, placeholder } = this.src
 
-        const loading =
-          (getOption('placeholder') && placeholder) ||
-          lazy.placeholder ||
-          options.placeholder
+        const loading = getOption('placeholder') ? placeholder : blank
 
         const blendIn = getOption('blendIn')
 
         return h('img', {
           attrs: {
             ...$attrs,
+            src,
+            srcset: loading,
             'data-srcset': srcSet,
-            width,
-            height
+            width
           },
-          class: { [styles.blendIn]: blendIn },
-          style: {
-            transition: `filter ${
-              blendIn
-                ? (typeof blendIn === 'number' && blendIn / 1000) || 0.5
-                : 0
-            }s`
-          },
-          directives: [
-            {
-              name: 'lazy',
-              value: {
-                src,
-                loading
-              }
-            }
-          ]
+          class: { [styles.blendIn]: blendIn }
         })
       }